+++ lib/models/room.dart
