+++ lib/widgets/enter_building_action_button.dart
@@ -1,9 +1,8 @@
 
 
+import 'package:concordi_around/provider/mapNotifier.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-
-import 'mapNotifier.dart';
 
 class EnterBuildingButton extends StatelessWidget{
