+++ lib/widgets/svg_floor_plans.dart
@@ -1,5 +1,5 @@
-import 'package:concordi_around/fadeIndexedStack.dart';
-import 'package:concordi_around/mapNotifier.dart';
+import 'package:concordi_around/provider/mapNotifier.dart';
+import 'package:concordi_around/widgets/custom/custom_fade_indexed_stack.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:provider/provider.dart';
